Added observer interfaces and outlines for wizard
The idea here is that each step will be able to react to any changes
made by other steps. The `izardSubject` object will contain the relevant
data for the whole process. At this stage the temporary storeage of
this data between requests is not decided. Possibly a cookie with a
serialized array of ids to reference.